+++ test/components/header/header.js
@@ -28,7 +28,7 @@ const initialState = {
   }
 };
 
-test('Header contains sessionTimeoutModal, correct number of nav items and excludes PDRs and Logs', function (t) {
+test('Header contains sessionTimeoutModal, networkErrorModal correct number of nav items and excludes PDRs and Logs', function (t) {
   const dispatch = () => {};
   const api = {
     authenticated: true
@@ -56,6 +56,9 @@ test('Header contains sessionTimeoutModal, correct number of nav items and exclu
 
   const modal = container.querySelector('.session-timeout-modal');
   t.truthy(modal);
+
+  const modalNetwork = container.querySelector('.network-error-modal');
+  t.truthy(modalNetwork);
 
   const navigation = container.querySelectorAll('nav li');
   t.is(navigation.length, 9);