+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - **CUMULUS-4003**
   - Remove D3 and related libraries from dashboard.
+- **CUMULUS-4114**
+  - Removed rules from pages affected by datepicker filter.
 
 ### Fixed